+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Note the first digit of every adapter version corresponds to the major version of the Chartboost Mediation SDK compatible with that adapter. 
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
+### 4.1.6.0.0
+- Adapter can now detect and report MobileFuseSDK initialization success/failure.
+- This version of the adapters has been certified with MobileFuseSDK 1.6.0.
+
 ### 4.1.4.4.1
 - Add support for Adaptive Banners.
 - This version of the adapters has been certified with MobileFuseSDK 1.4.4.